requirements.txt
openai>=1.0.0
langchain>=0.0.267
langchain-openai>=0.0.1
python-dotenv>=1.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
tiktoken>=0.5.1
pandas>=2.0.0
tqdm>=4.66.1
tenacity>=8.2.3
aiohttp>=3.8.5
pydantic>=2.4.2
asyncio>=3.4.3 